Updated openstack/openstack Project: openstack/tempest  3e3942e11c963da056d21e3061221aac9b60f577
Remove set_nic_state from old remote_client

set_nic_state was just one-line command wrapper method and it was
not worth to keep it as a common method. That made stacktrace
deep unnecessarily.
This patch removes set_nic_state for the cleanup.

NOTE: This method is not used at outside of Tempest.

Change-Id: Ida09a3b4098980c1644cc096453c34ef6e1c2a37