Updated openstack/openstack Project: openstack/tempest  4f78af83bf85646c554e1ef05ff3b037570ecdf9
Remove assign_static_ip from old remote_client

assign_static_ip was just one-line command wrapper method and it
was not worth to keep it as a common method. That made stacktrace
deep unnecessarily.
This patch removes assign_static_ip for the cleanup.

NOTE: This method is used in Congress, it is better to switch it
      on Congress side before this patch. The congress patch is
      Ie527ca8f163faa69792c78791878feee9f1e3956

Depends-On: Ie527ca8f163faa69792c78791878feee9f1e3956
Change-Id: Ie03d54f876b10b2546a1406ac944b4d09f82b510