Updated openstack/openstack Project: openstack/tempest  22159e9bb36550a369ec1ccc41c2a0cb183515ca
Nova: Move _Websocket class to a common place

To avoid to introduce a dependency to "websocket-client", we re-use
the internal "_Websocket" class which encapsulates the sockets.
This class got introduced with commit 1f87a5611 for the VNC console
test cases.
This change moves this class to a common place, so that it can be
used by VNC and (in a later change) "serial console" test cases
equally.

Change-Id: I79065865c876549269b2bb962a6826ac2fd0a403