Updated openstack/openstack Project: openstack/api-wg  ddc2f166dd76902e13eceaf0ecff5905214e1391
Clarify the meaning of BODY

http.rst contains "HAS BODY?", but it is not so clear that it means
a request body or a response body. On the bug report 1677360, we said
HEAD can be used as low-cost version of GET by none response body.
On the table, both of HEAD and GET is NO on "HAS BODY?". So the column
means a request body. This patch makes the meaning clear.

Change-Id: Idbfc6185ffe33774ce89a591cb034288e0953a36
Related-Bug: #1677360