Updated openstack/openstack Project: openstack/designate  7f4a9a65ea0c102f8908c33ddc98a95c825b6fad
Add SOA to RECORD_TYPES in sqlalchemy.tables and remove eventlet constraint

sqlalchemy 1.1 is more strict on enum values. Because of that, the fact
that the two RECORD_TYPES enums are out of sync makes things sad.

This patch also pin eventlet version to 0.19.0 since eventlet >=0.20.0 in
upper constraint still make UT fail.

Co-Authored-By: Hieu LE <hieulq@vn.fujitsu.com>
Change-Id: I4de68749e9eddbddbc7f8d8df68ce3ea71cc4fcb