Updated openstack/openstack Project: openstack/requirements  c579109a9224a636e4f94100e796b1b4d03e9def
Add munch

munch is a lightweight library and is what shade returns from its
API instead of dicts of custom objects. In order for shade to sync
with global-requirements now that it's in the tent, we need to add them.

- Is the library actively maintained?

Yes. munch itself is a fork of bunch, which was not maintained. The most
recent release was made 9 days ago, and it has been in use in shade
with good results since Oct 19, 2015.

- Is the library good code?

Yes. It does everything one wants, and nothing one doesn't want.

- Is the library python 3 compatible?

Yes.

- Is the library license compatible?

Yes, it is MIT licensed.

- Is the library already packaged in the distros we target (Ubuntu
  latest / Fedora latest)?

Yes, in ubuntu since Xenial, in Fedora since at least F22.

- Is the function of this library already covered by other libraries
  in ``global-requirements.txt``?

No, it's not.

- Is the library required for OpenStack project or related dev or
  infrastructure setup? (Answer to this should be Yes, of course)
  Which?

Shade uses munch objects as part of its public API.

Change-Id: I4dd6be7a31d16a430c48ed4738fac4769c65975a