Updated openstack/openstack Project: openstack/nova  07fcf70cae2f91bef4e1be16fe1244c288080277
doc: add some documentation around quotas

There are always some questions about how quotas work
in Nova, particularly how they are checked against
the configuration vs the 'default' quota class in the
database, and global vs project-specific quotas.

This change starts a document on quotas a high-level.
It does not get into the lower-level details of how
quotas work or are designed in Nova, or the issues
that can arise with the existing quota system or future
plans.

Change-Id: Iabe3d40f44c59f14fa89ceddc556497d2885845c