Updated openstack/openstack Project: openstack-infra/shade  c80225418701e3d4e6a688c4ff308332faa47711
Add bare parameter to get/list/search server

The bare parameter tells shade to skip all additional calls it makes to
fill in extra information on the server record from other services, such
as fixing the addresses dict. This is useful internally in shade where
there are times when one is getting the server dict just to get its id
so other operations, such as getting the console log or attaching
security groups, can be performed.

Change-Id: Ic9cc6e8be4f4e30f76d3abb655d9a3cf0cb9918b