Updated openstack/openstack Project: openstack/neutron  dc84e891ca4aa13e35d841d12ae791271e39b1e3
Skip DHCP agent query in provisioning block setup

Provisioning blocks merged in Newton so for Pike we can
safely assume we are not running with Liberty agents that
don't notify the server when the port is ready.

This allows us to skip a query to the agents and configuration
parsing on every port provisioning block setup.

Related-Bug: #1453350
Change-Id: I8111469ad4b0d88580bff7a77492ad95af8e9377