Updated openstack/openstack Project: openstack/nova  fd3ee2a3ccc60f173716d8d4ed8fb15ad709d23d
Do not attempt to load osinfo if we do not have os_distro

We get a warning logged every time we try to load up osinfo
with an image metadata that does not have the 'os_distro'
property set. We should be smarter and just not try to load
osinfo at all if we know we cannot get results.

Change-Id: I79ebea3ed222fb0db01d0a33c62a677d0381f66c
Closes-Bug: #1675602