Updated openstack/openstack Project: openstack/openstack-manuals  1076fbd9f3888b8140c3d8a149e3afd6f1b5daa8
Revert "[install-guide] Workaround step for Dashboard"

This reverts commit 5d7e6cd9e7be8e70744576af749cb454965f7c93.

This patch isn't really needed; we have a warning in Horizon
(https://review.openstack.org/#/c/443716/) which has been proposed for
backport to stable. This is just a bug in the Ubuntu packages, and docs
shouldn't be expected to carry a workaround for a bug which already
logs a warning.

Change-Id: I3888a7cd671394d67d388106defb2358389f13e8