Updated openstack/openstack Project: openstack-infra/project-config  4fc0a4baca257830d6cb238332167c90dc6af6db
Add requirements gate jobs to shade and requestsexceptions

Shade is an OpenStack project now, so should be tracking
global-requirements. This depends on the patches to add jmespath and
munch to global-requirements so that shade isn't left unable to land
patches. Also, remove infra-publish jobs from requestexceptions - they
were not used anyway.

Change-Id: I334c4870e6bb608682573e69fff727264462a1e1
Depends-On: I963a558f5e294521ad7b067ae95cc02d379cf73b