Updated openstack/openstack Project: openstack-infra/shade  49095d5323edcbb0303f74d8f6f44e44e8b48594
Change metadata to align with team affiliation

Shade is its own team now, update "author" to be OpenStack.

Note, the homepage setting does not exist until we tag a release after
the depends-on project-config patch lands. But this metadata also will
not go to PyPI until we tag a release - thus the depends-on.

Depends-On: Id3a087a41491eb8bf8eb51a0247183c6b91a297a
Change-Id: I4fc7ed2ecca0d218be7f1e1e34dcfa4c4b623d4e