Updated openstack/openstack Project: openstack/governance  13c619b9b15fca197e7180003493c2cae7b3df46
More gracefully handle intra-openstack moves

Splitting a project out into its own team causes problems with the
diversity check if the new project is inside of another project. This is
because there is no {project}-group stackalytics container, only a
{{project}} one.

Trap for the 404 error and re-point to {{project}} if {{project}}-group
is not found.

Change-Id: I68b1b7c4d98810f1d710351e4477fb83ef1fad21