Updated openstack/openstack Project: openstack-infra/shade  235d097d65fd42373ade436c5624fb05f67f546b
Fix a few minor annoyances that snuck in

We had the normal infra pep8 ignores marked, but follow some of them by hand
in this codebase anyway. While E12* are not valid pep8, the prevailing style
of shade is in agreement with them. While we weren't enforcing them, a few
instances snuck in. Fix them and just turn the flags on.

Change-Id: Iea2bd9a99eb8b20dd40418ac8479b809d3872de8