Updated openstack/openstack Project: openstack-infra/shade  03e9e1ff2dba02098851d8157f83b0f1e3110f28
Enable H201 - don't throw bare exceptions

I agree - bare except is bad - it can cause ctrl-c to stop working,
which is totally evil. (I mean, unless that is the intent, in which case
it's the right thing to do. None of these cases are cases when we desire
ctrl-c to stop working)

Change-Id: Ieffa26e8da77199c9ef7238a6c9f0704fc581ba6