Updated openstack/openstack Project: openstack/keystone  539b6e8074413c108c44544454ae5ab354beb23d
Remove policy file from source and refactor tests

This patch removes the policy file now that policy in code has merged
and we now have all the default policy from policy.json in code. In
addition it refactors a couple of tests to use the in code policy
instead of the policy.json file when testing.

Depends-On: Ib1d9a51a78e2a84a3d7294dc8782605a681fa9e8
Change-Id: I23136627cb095015352b8eb09c31b2b2f1ce315f
Implements: blueprint policy-in-code