Updated openstack/openstack Project: openstack/cinder  d6882838af2c2c55e39a7060402137fa627986fa
Add an optional db session argument to cinder.db.sqlalchemy.api.
quota_allocated_get_all_by_project(). Then in quota_reserve()
pass the session that is fetched for that method so that a new
session does not have to be made. This can avoid DB timeouts
and retries when a lot of concurrent operations are hitting
quota_reserve().

Added session to method mocking in the UT cases and removed a
mocking instance that was dead code since it was already mocked
in the specific tests needed.

Change-Id: I43b5035926eb8b65e7913ac0a8c58e6b9ebb5619
Closes-Bug: #1676574