Updated openstack/openstack Project: openstack-infra/tripleo-ci  b221ac3264901ccbd5f958ada972380180fc7d04
Use shorter timeout for update step

The update step should take less time than the create step, and if
we use the full create timeout then we will run past the gate
timeout and get killed with no logs.  Based on the current graphite
metrics, all successful jobs are finishing update in under 45
minutes so that should be safe value.

Change-Id: I0e9761c81b95fdf85c4f57daedb630a6e0cb9db2
Closes-Bug: 1674770