Updated openstack/openstack Project: openstack/governance  34ba24fae986815a70e47663e761939b7605b2f5
Move castellan under Oslo governance

The Barbican team would like to move Castellan under Oslo. The hope is
to send a clear signal to everyone that castellan is the preferred
library for their key manager needs and sufficiently abstracts hard
dependency to barbican and attract new folks with alternate implementations.

http://lists.openstack.org/pipermail/openstack-dev/2017-March/114335.html
http://eavesdrop.openstack.org/meetings/barbican/2017/barbican.2017-03-20-20.00.txt

Change-Id: I571dc981a198fcde1a8bdf626ce159b860d452aa