Updated openstack/openstack Project: openstack/neutron  830fe9802219c3b16d7ca76ffa8ee3650e832b9e
Handle empty body in add_router_interface

when passing an empty dict body {} to add_router_interface[1], internal
server error will be raised.
There is argument check in the cli which checks either subnet or port id
is needed, but not in the server side when the body is empty.

[1]:
https://developer.openstack.org/api-ref/networking/v2/?expanded=add-interface-to-router-detail

Change-Id: Iad48dadb1235bc45787a1e87ff3f0d7a21df18af
Closes-bug: #1674069