Updated openstack/openstack Project: openstack/nova  88bc8dc5ce32748452c9d3acda9f35e77fedb6ce
api-ref: fix description of volumeAttachment for attach/swap-volume

The description of the volumeAttachment request parameter was actually
the description of the response parameter, which included more fields
than we allow on the POST and PUT requests for os-volume_attachments.

This fixes the descriptions for both POST and PUT and also includes a
tiny fix for the wording on the existing volumeAttachment parameter.

Change-Id: I4ccd4ac12e24b232925875fdb5fb568c2bfaf417
Closes-Bug: #1675536