Updated openstack/openstack Project: openstack/keystone  39c6b0ff537b822486d4fc187159464ea04bf912
Make use of Dict-base including extras explicit

This change replaces the use of DictBase with the ModelDictMixin
for any SQL models that do not contain an extra column and renames
the DictBase to a more descriptive name of ModelDictMixinWithExtras.

A Docstring has been added indicating the continued usage of
ModelDictMixinWithExtras should not be done for any "new"
models.

Change-Id: I9a4767cacf7620e878df70084060f3e43e1318df