Updated openstack/openstack Project: openstack-infra/os-loganalyze  5b7ad194dd1401eaca4952b806d374379a1204b7
Remove swift support

The swift support was never moved into production for infra because of
an unsolved issue around index files, so the logs in swift approach
was ended. This takes out all the swift supporting code to drop the
complexity of the code base. As this is very intermittently maintained
tool, less code makes any future work easier to do.

Change-Id: Iabc4cffc35633fd808556ebb68a8bdd5a5d1fde0