Updated openstack/openstack Project: openstack/keystone  b4164f158e36bea48b9b945af5030d4c97091ed9
Move revoke events to DocumentedRuleDefault

The overall goal is to define a richer policy for deployers
and operators[0]. To achieve that goal a new policy
class was introduce that requires additional parameters
when defining policy objects.

This patch switches our revoke events policy object to
the policy.DocumentedRuleDefault and fills the
required policy parameters as needed.

[0] I2b59f92545c5ead2a883d358f72f3ad3b3dfe1a6

Change-Id: Idfa3e5cd373c560035d03dfdef4ea303e28a92fc
Partially-Implements: bp policy-docs