Updated openstack/openstack Project: openstack/neutron  e0dd9158210785ff0251f4c883549f6c33e82929
Remove stale floating IP addresses from rfp devices

Old versions of the DVR code used to configure floating
IP addresses directly on the rfp-* devices in the
qrouter namespace.  This was changed in Mitaka to use
routes, but these stale floating IP addresses were
never cleaned-up when the l3-agent was restarted.

Add a check for these addresses and remove them at
startup if they exist.  This can be removed in a cycle
since once they are cleaned they will not be added
back by the agent.

Change-Id: Id512d213cd7ee11da913a4e4b0da20c3ad5420b0
Closes-bug: #1675187