Updated openstack/openstack Project: openstack/python-novaclient  0010ab79b45cb111634c83bd7415752984a8d767
Update Compute API Guide pointer

Since the description of the individual API calls have been moved to
the API Reference, the API Guide is more or less what the "overview"
in the original guide was.  Therefore I just pointed to the guide, and
removed the phrase "-- the overview, at least --".

Change-Id: I8b1cb6a7ea4469a91f0dfe69d26f4ffec178ed43
Closes-Bug: 1678591
Signed-off-by: Simon Leinen <simon.leinen@gmail.com>