Updated openstack/openstack Project: openstack-infra/shade  aa58173f0d60a8f04318b51ff6d1490e7ee10cfb
Add ability to configure extra_specs to be off

We fetch flavor extra_specs for each flavor for consistency with old
behavior. However, it's costly and often just simply not needed.

Query clouds.yaml for extra client settings and if "get_flavor_extra_specs" is
set to false, set the default for list_flavors to false.

Change-Id: Iaea1c41e8e0ae504cb080d7d37407de12be96fd1