Updated openstack/openstack Project: openstack-infra/project-config  32f426a2b9ecea7a92a78cb9f1fa509c0371f966
Tweak hooks for golang dsvm job

$BASE is not yet defined when the gate hooks are
registered, so we cannot check for the existence
of those files. Instead we check the existence of
files when the hooks are run. Also if the gate
hook is not defined then we do what is usually
done which is to run the devstack-vm-gate.sh script

Change-Id: I8e664f6633289a8514834598d72f82cc403497f4