Updated openstack/openstack Project: openstack-infra/project-config  7f0f18ce5be67963d09d6b8c6483ee4dc63db0be
[Manila] Add 2 jobs for manila-image-elements project

For the moment, we have jobs that just build images for this
project, but they are not tested for usage with appropriate drivers.

So, add two tempest jobs, that will use built image with proposed
changes to "manila-image-elements" project and use it for manila
'generic' and 'container' (docker) drivers.

Add those jobs only to check pipeline and make them non-voting.

Change-Id: I3698044bbcaa94bddcfa192841520b4d7ea32016