Updated openstack/openstack Project: openstack/ironic  41bd6f28d643e5c779b67ca13c81c1416aa586c0
[Devstack] Rework VMs connection logic

Devstack currently plugs the simulated baremetal VMs into OVS using the
libvirt bridge driver, this caused a problem because libvirt unplugs the
VM from the network when it is turned off. To fix this an extra bridge
was added between the VM and OVS to allow the OVS port to persist even
when the VM was turned off. This patch replaces how the devstack
simulated baremetal VMs are plugged into OVS with a manually created tap
interface, this removes the need for an extra bridge, because manually
created tap interfaces aren't unplugged when the VM is turned off.

Allow to connect several interfaces to a node by setting
IRONIC_VM_INTERFACE_COUNT devstack variable.

Co-Authored-By: Vasyl Saienko <vsaienko@mirantis.com>

Change-Id: Iafd470445d59f0e2009e65ddaf65a6c603a1e1c1