Updated openstack/openstack Project: openstack/nova  a24ce5b12776c854a473b4fb6a9f3c27731d3af9
Distinguish between cells v1 and v2 in upgrades doc

The note about rolling upgrades not being supported for cells
was referring to cells v1, so let's make that clear.

We will support rolling upgrades for cells v2 deployments, we
just haven't documented it yet (or done it in a grenade job).

Change-Id: I6a8b3d3aaa49859e8b769a296684ac96883719f8