Updated openstack/openstack Project: openstack/python-openstackclient  279066b67c536a2970446779b41410315647ecaf
Change noauth strategy for plugin loading

Don't do it.

os-client-config's plugin loading has been causing this pain for a long
time, removing the KSC hack-around in o-c-c unmasked this again.  So when
auth is not reuired, just don't let o-c-c do any plugin loading at all.

Ever.

Of course, this shouldn't be in OSC either, but we have to do this until
the equivalent fix lands in osc-lib, is released and makes it into the
global requirements minimum version.

Depends-on: Ie68c82f7b073012685f0513b615ab1bf00bc0c3a
Change-Id: Ifdf65f3bb90fb923947a2cbe80a881d71a3fee56