Updated openstack/openstack Project: openstack/openstack-manuals  5ecf993a6bea5d1cfdb3a77c7684d146cdd9e5e9
Add Identity v3 API varibales to the designate example section.

The example provided for the designate section in neutron.conf is missing
keystone v3 parameters. If the OpenStack deployment is using keystone
v3 API then the current set of variables as shown in the example will
not be enough and neutron-server will complain about the missing keystone
variables.

Change-Id: Ic9a386f50812f89903f9b35fdfddb6e9850583fd