Updated openstack/openstack Project: openstack/tempest  82e5b2d8374ff57ef7e45cc596f447fe264d8426
Rename server arg in wait_for_interface_status method

This is very minor, but the server formal argument to the
wait_for_interface_status method is actually a server id.

This changes the name of the parameter to be clear on the
type of that object, since it could easily be confused
with a full server response body dict.

Change-Id: Id7cab21ab75092226f0784877ae55f17e880229c