Updated openstack/openstack Project: openstack/cinder  96d3f33215007456d8eeb7f36a8b3f8c923bd735
Add more specific error message

During restoring backup without providing volume_id if we have
error in volume creation process we will have error message
'Volume to be restored must be available' that is ambiguous.
This message it's more connected with that situation when we
provide volume_id for restoring operation and it isn't in
available state.

Add more specific error message for this situation.

Change-Id: Iaf2d3cc942e6f6492dd3270e188ffa6676ac635a