Updated openstack/openstack Project: openstack/tempest  3f9b3b1191ab3b7e91bd1525f10a1c202009c4c3
Add tempest test writing guide

This commit is the start of a new doc guide for helping people figure
out how to write good tempest tests. This is just a starting point for
the document and doesn't cover every aspect of writing a test case in
tempest. Right now it just covers the setupClass phase for writing
TestCases. It will be expanded upon in future commits to give a complete
guide to writing tests in Tempest.

Change-Id: Ie0db2079163f536ae72668c8fc9129ae4c6389e2