Updated openstack/openstack Project: openstack/barbican  4f8ee048ea46bb2ff2fc7e57067ac51068cb8204
Fix install guide docs for SUSE

First, the example package install command neglected to include any of
the package names, so this patch adds the correct package names as well
as omits the non-essential, personal-preference command flags so as to
be consistent with the examples in the main OpenStack install guide.

Second, like the barbican packages in RDO, the OBS barbican packages do
not include an openstack-barbican-api systemd unit file and instead
provide an Apache vhost sample file, so this patch modifies the service
instructions to configure the service with Apache.

Change-Id: I13bfeef4f06602984d2ffb727212ab6d39797401