Updated openstack/openstack Project: openstack/nova-specs  b640fea68e28dcac4dc30539f11687243f485a04
Add project/user association to placement

This cycle we are changing the quota system to count resources to check
quota instead of tracking usage and reservations separately. As things
currently stand, we must query cell tables to count things like cores
and ram to check against quota limits. There are a couple of problems
with the current approach:

  1. Querying all cells for instances owned by a project and summing their
     cores and ram counts is not efficient.
  2. Quota usage becomes effectively "freed" if contact with one or more
     cells is lost for any reason, until the cells return.

To address these problems, we propose adding project and user associations
to placement for consumers.

Spec for blueprint placement-project-user

Change-Id: I340d403a6640661dd3a5eacf2a4145071396afe9